Fix flex display and properties for header and left/right elements
The header was not correctly having flex properties assigned because the
rule was selecting the header type, not the header class as the header
on this page is a div and not the header type.

Right and left divs were assigned to have basis as specified by the
explicit dimensioning, and were assigned not to shrink so narrowing the
viewport does not cause the texts for elements to bunch up/wrap.