Fix header and close button alignment by placing in container
Header and close button are supposed to be aligned so placed them in a
flex container together.

Also fix alignment of the flex items in the text.

Need to figure out why the text does not display the same as in the
desired outcome image - doesn't extend far right out enough.